Add CNode.insertIf()
Break about INode.insertIf() for CNodes, inlining INode.insertDual()
into its sole caller.
